+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - `search`
   - Support alternate names for 4 color combos (#24)
   - Fix issue where some values from the database had extra spaces or line breaks
+- `autocomplete`
+  - Add support for full names of 4 color combos
 
 # 0.12.0