+++ CHANGELOG.md
@@ -1,7 +1,22 @@
 # unreleased
 
+- `search`
+  - Support alternate names for 4 color combos (#24)
+  - Fix issue where some values from the database had extra spaces or line breaks
+- `autocomplete`
+  - Add support for full names of 4 color names
+  - Add support for alternate 4 color names
+
+# 0.12.0
+
 - `autocomplete`
   - filter out repeating values that differ only in casing and punctuation
+- `search`
+  _Breaking Changes_
+
+  - Use `spellbookId` in place of `id` param (#21)
+  - Add `id`, `ids`, `c` as aliases for `coloridentity` (#23)
+  - Prevent passing values that normalize to empty strings (#26)
 
 # 0.11.0